# internal/api/api.yaml
openapi: 3.0.0
info:
  title: Payment Sandbox API
  version: 0.1.0
  description: Authentication APIs for merchant onboarding and login.
...
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://staging.api.example.com
    description: Staging server
  - url: https://api.example.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /pay/{token}:
    get:
      operationId: ResolvePaymentLink
      summary: Resolve public payment link
      tags: [Payment Public]
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Payment link resolved
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pay/{token}/intents:
    post:
      operationId: CreatePaymentIntent
      summary: Create payment intent
      tags: [Payment Public]
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentIntentRequest"
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/refunds:
    post:
      operationId: RequestRefund
      summary: Merchant request refund
      tags: [Merchant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestRefundRequest"
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: ListRefundHistory
      summary: Refund history
      tags: [Merchant]
      parameters:
        - in: query
          name: merchantId
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        "200":
          description: Paginated refund history
...
  /merchant/topups:
    post:
      operationId: RequestTopUp
      summary: Merchant request top-up
      tags: [Merchant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestTopUpRequest"
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: ListTopUpHistory
      summary: Top-up history
      tags: [Merchant]
      parameters:
        - in: query
          name: merchantId
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        "200":
          description: Paginated top-up history
...
  /admin/dashboard/stats:
    get:
      operationId: GetDashboardStats
      summary: Admin dashboard statistics
      tags: [Admin]
      parameters:
        - in: query
          name: merchantId
          schema: { type: string }
        - in: query
          name: startDate
          schema: { type: string, format: date-time }
        - in: query
          name: endDate
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Dashboard statistics
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/payment-intents/{intentId}/simulate:
    post:
      operationId: SimulatePaymentIntent
      summary: Admin simulate payment intent outcome
      tags: [Admin]
      parameters:
        - in: path
          name: intentId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulatePaymentIntentRequest"
      responses:
        "200":
          description: Simulation applied
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/refunds/{refundId}/approve:
    post:
      operationId: ApproveRefund
      summary: Admin approve refund
      tags: [Admin]
      parameters:
        - in: path
          name: refundId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Refund approved
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/refunds/{refundId}/reject:
    post:
      operationId: RejectRefund
      summary: Admin reject refund
      tags: [Admin]
      parameters:
        - in: path
          name: refundId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Refund rejected
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/refunds/{refundId}/process:
    post:
      operationId: ProcessRefund
      summary: Admin process refund outcome
      tags: [Admin]
      parameters:
        - in: path
          name: refundId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRefundRequest"
      responses:
        "200":
          description: Refund processed
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/topups/{topupId}/status:
    post:
      operationId: UpdateTopUpStatus
      summary: Admin update top-up status
      tags: [Admin]
      parameters:
        - in: path
          name: topupId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTopUpStatusRequest"
      responses:
        "200":
          description: Top-up updated
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices:
    post:
      operationId: CreateInvoice
      summary: Create invoice
      tags: [Invoice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: ListInvoices
      summary: List invoices
      tags: [Invoice]
      parameters:
        - in: query
          name: merchantId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: string
        - in: query
          name: pageSize
          schema:
            type: string
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/{id}:
    get:
      operationId: GetInvoice
      summary: Get invoice detail
      tags: [Invoice]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: UpdateInvoice
      summary: Update invoice
      tags: [Invoice]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: DeleteInvoice
      summary: Delete invoice
      tags: [Invoice]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Invoice deleted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: Refresh
      summary: Refresh JWT token
      tags: [Auth]
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid auth context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreatePaymentIntentRequest:
      type: object
      required: [method]
      properties:
        method:
          type: string
          description: WALLET, VA_DUMMY, or EWALLET_DUMMY
    RequestRefundRequest:
      type: object
      required: [invoiceId, merchantId, amount]
      properties:
        invoiceId:
          type: string
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
    RequestTopUpRequest:
      type: object
      required: [merchantId, amount]
      properties:
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
        requestKey:
          type: string
    CreateInvoiceRequest:
      type: object
      required: [merchantId, amount]
      properties:
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
        currency:
          type: string
        dueDate:
          type: string
          format: date-time
    UpdateInvoiceRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          format: int64
        currency:
          type: string
        dueDate:
          type: string
          format: date-time
    InvoiceResponse:
      type: object
      required: [id, invoiceNumber, merchantId, amount, currency, status, paymentToken, createdAt]
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
        currency:
          type: string
        status:
          type: string
        paymentToken:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    InvoiceListResponse:
      type: object
      required: [items, page, size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceResponse"
        page:
          type: integer
        size:
          type: integer
...
    AuthResponse:
      type: object
      required:
        - token
        - expiresAt
        - userId
        - email
        - role
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        userId:
          type: string
        email:
          type: string
        role:
          type: string
    SimulatePaymentIntentRequest:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          description: SUCCESS or FAILED
    ProcessRefundRequest:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    UpdateTopUpStatusRequest:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Detailed error message